requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.1
pdf2image>=1.16.3
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
opencv-python>=4.8.0
scikit-learn>=1.3.0
datasets>=2.14.0
huggingface-hub>=0.19.0
streamlit>=1.28.0
chromadb>=0.4.18
openai>=1.0.0
PyYAML>=6.0
tqdm>=4.66.0
